+++ google-beta/services/appengine/resource_app_engine_standard_app_version_test.go
@@ -11,6 +11,7 @@ import (
 )
 
 func TestAccAppEngineStandardAppVersion_update(t *testing.T) {
+	t.Skip("https://github.com/hashicorp/terraform-provider-google/issues/18936")
 	t.Parallel()
 
 	context := map[string]interface{}{
